# Conflicts:
#	.gitignore
#	README.md
#	src/IMS/Controller/Command.java
#	src/IMS/Controller/Controller.java
#	src/IMS/Controller/ControllerImpl.java
#	src/IMS/IMSException.java
#	src/IMS/IMSUtil.java
#	src/IMS/Model/Category.java
#	src/IMS/Model/Customer.java
#	src/IMS/Model/DatabaseUtil.java
#	src/IMS/Model/DummyModel.java
#	src/IMS/Model/InventoryModel.java
#	src/IMS/Model/InventoryModelImpl.java
#	src/IMS/Model/Item.java
#	src/IMS/Model/RetailStore.java
#	src/IMS/Model/Sale.java
#	src/IMS/Model/Status.java
#	src/IMS/Model/SupplyOrder.java
#	src/IMS/Model/Vendor.java

+++ src/IMS/Model/Item.java
@@ -121,7 +121,8 @@ public class Item {
         return itemPrice;
     }
 
-    public void setItemPrice(int itemPrice) {
+    public void setItemPrice(double itemPrice) {
         this.itemPrice = itemPrice;
     }
 }
+
